move repeat code to utils.py
